+++ semi/src/semi/beans/board/BoardTextDao.java
@@ -79,7 +79,6 @@ public class BoardTextDao {
 		con.close();
 		return boardtextdto;	
 	}
-
 	//편집 메소드
 		public void btedit(BoardTextDto dto) throws Exception{
 			Connection con = getConnection();
@@ -96,4 +95,3 @@ public class BoardTextDao {
 			con.close();
 		}
 }
-
