--- semi/src/src/semi/beans/ba_board/BA_BoardDto.java
@@ -1,81 +1,0 @@
-package semi.beans.ba_board;
-
-public class BA_BoardDto {
-	private int board_no;
-	private String writer;
-	private String title;
-	private String content;
-	private String wdate;
-	private String udate;
-	private int up;
-	private int down;
-
-	public BA_BoardDto() {
-		super();
-	}
-
-	public int getBoard_no() {
-		return board_no;
-	}
-
-	public void setBoard_no(int board_no) {
-		this.board_no = board_no;
-	}
-
-	public String getWriter() {
-		return writer;
-	}
-
-	public void setWriter(String writer) {
-		this.writer = writer;
-	}
-
-	public String getTitle() {
-		return title;
-	}
-
-	public void setTitle(String title) {
-		this.title = title;
-	}
-
-	public String getContent() {
-		return content;
-	}
-
-	public void setContent(String content) {
-		this.content = content;
-	}
-
-	public String getWdate() {
-		return wdate;
-	}
-
-	public void setWdate(String wdate) {
-		this.wdate = wdate;
-	}
-
-	public String getUdate() {
-		return udate;
-	}
-
-	public void setUdate(String udate) {
-		this.udate = udate;
-	}
-
-	public int getUp() {
-		return up;
-	}
-
-	public void setUp(int up) {
-		this.up = up;
-	}
-
-	public int getDown() {
-		return down;
-	}
-
-	public void setDown(int down) {
-		this.down = down;
-	}
-	
-}
